new version warning banner
Read the Docs no longer injects this information during the build. It
also required a new build of each old version whenever a new version was
released, which is not practical. Instead, use JavaScript to determine
the banner when viewed. Query PyPI for the list of versions, and compare
the docs version. Use CSS for scroll offset.